Rename "V2" namespace to "Services"
