+++ src/GW2NET/Factories/Services/AuthorizedServices.cs
