+++ src/GW2NET/Factories/Services/ColorRepositoryFactory.cs
