+++ src/GW2NET/Factories/Services/TrinketConverterFactory.cs
