+++ src/GW2NET/Factories/Services/UpgradeComponentConverterFactory.cs
