+++ src/GW2NET/GW2NET.csproj
@@ -129,36 +129,36 @@
     <Compile Include="Factories\FactoryForLocal.cs" />
     <Compile Include="Factories\FactoryForRendering.cs" />
     <Compile Include="Factories\RepositoryFactoryBase.cs" />
-    <Compile Include="Factories\V2\AccountRepositoryFactory.cs" />
-    <Compile Include="Factories\V2\ArmorConverterFactory.cs" />
-    <Compile Include="Factories\V2\ArmorSkinConverterFactory.cs" />
-    <Compile Include="Factories\V2\CharacterRepositoryFactory.cs" />
-    <Compile Include="Factories\V2\ColorRepositoryFactory.cs" />
-    <Compile Include="Factories\V2\CombatAttributeConverterFactory.cs" />
-    <Compile Include="Factories\V2\ConsumableConverterFactory.cs" />
-    <Compile Include="Factories\V2\ContainerConverterFactory.cs" />
-    <Compile Include="Factories\V2\ContinentRepositoryFactory.cs" />
-    <Compile Include="Factories\V2\Services.cs" />
-    <Compile Include="Factories\V2\AuthorizedServices.cs" />
-    <Compile Include="Factories\V2\FactoryForV2Commerce.cs" />
-    <Compile Include="Factories\V2\FileRepositoryFactory.cs" />
-    <Compile Include="Factories\V2\GatheringToolConverterFactory.cs" />
-    <Compile Include="Factories\V2\GatheringToolSkinConverterFactory.cs" />
-    <Compile Include="Factories\V2\GizmoConverterFactory.cs" />
-    <Compile Include="Factories\V2\ItemConverterFactory.cs" />
-    <Compile Include="Factories\V2\ItemRepositoryFactory.cs" />
-    <Compile Include="Factories\V2\MapsRepositoryFactory.cs" />
-    <Compile Include="Factories\V2\RecipeConverterFactory.cs" />
-    <Compile Include="Factories\V2\RecipeRepositoryFactory.cs" />
-    <Compile Include="Factories\V2\SkinConverterFactory.cs" />
-    <Compile Include="Factories\V2\SkinRepositoryFactory.cs" />
-    <Compile Include="Factories\V2\ToolConverterFactory.cs" />
-    <Compile Include="Factories\V2\TrinketConverterFactory.cs" />
-    <Compile Include="Factories\V2\UnlockerConverterFactory.cs" />
-    <Compile Include="Factories\V2\UpgradeComponentConverterFactory.cs" />
-    <Compile Include="Factories\V2\WeaponConverterFactory.cs" />
-    <Compile Include="Factories\V2\WeaponSkinConverterFactory.cs" />
-    <Compile Include="Factories\V2\WorldRepositoryFactory.cs" />
+    <Compile Include="Factories\Services\AccountRepositoryFactory.cs" />
+    <Compile Include="Factories\Services\ArmorConverterFactory.cs" />
+    <Compile Include="Factories\Services\ArmorSkinConverterFactory.cs" />
+    <Compile Include="Factories\Services\CharacterRepositoryFactory.cs" />
+    <Compile Include="Factories\Services\ColorRepositoryFactory.cs" />
+    <Compile Include="Factories\Services\CombatAttributeConverterFactory.cs" />
+    <Compile Include="Factories\Services\ConsumableConverterFactory.cs" />
+    <Compile Include="Factories\Services\ContainerConverterFactory.cs" />
+    <Compile Include="Factories\Services\ContinentRepositoryFactory.cs" />
+    <Compile Include="Factories\Services\Services.cs" />
+    <Compile Include="Factories\Services\AuthorizedServices.cs" />
+    <Compile Include="Factories\Services\FactoryForV2Commerce.cs" />
+    <Compile Include="Factories\Services\FileRepositoryFactory.cs" />
+    <Compile Include="Factories\Services\GatheringToolConverterFactory.cs" />
+    <Compile Include="Factories\Services\GatheringToolSkinConverterFactory.cs" />
+    <Compile Include="Factories\Services\GizmoConverterFactory.cs" />
+    <Compile Include="Factories\Services\ItemConverterFactory.cs" />
+    <Compile Include="Factories\Services\ItemRepositoryFactory.cs" />
+    <Compile Include="Factories\Services\MapsRepositoryFactory.cs" />
+    <Compile Include="Factories\Services\RecipeConverterFactory.cs" />
+    <Compile Include="Factories\Services\RecipeRepositoryFactory.cs" />
+    <Compile Include="Factories\Services\SkinConverterFactory.cs" />
+    <Compile Include="Factories\Services\SkinRepositoryFactory.cs" />
+    <Compile Include="Factories\Services\ToolConverterFactory.cs" />
+    <Compile Include="Factories\Services\TrinketConverterFactory.cs" />
+    <Compile Include="Factories\Services\UnlockerConverterFactory.cs" />
+    <Compile Include="Factories\Services\UpgradeComponentConverterFactory.cs" />
+    <Compile Include="Factories\Services\WeaponConverterFactory.cs" />
+    <Compile Include="Factories\Services\WeaponSkinConverterFactory.cs" />
+    <Compile Include="Factories\Services\WorldRepositoryFactory.cs" />
     <Compile Include="GW2Bootstrapper.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="$(SolutionDir)SharedAssemblyInfo.cs">
